Implement a standard way of including icons
