+++ packages/common/package.json
@@ -8,6 +8,7 @@
   "author": "Eldarion",
   "license": "MIT",
   "dependencies": {
+    "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/vue-fontawesome": "^0.1.8",
     "lodash.debounce": "^4.0.8"
