+++ packages/reader-image-mode/package.json
@@ -8,6 +8,7 @@
   "author": "Eldarion",
   "license": "MIT",
   "dependencies": {
+    "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "@scaife-viewer/common": "^0.0.2",
     "@scaife-viewer/widget-reader": "^0.0.5",
     "graphql-tag": "^2.10.3",
