+++ packages/reader-image-mode/src/index.js
@@ -1,3 +1,4 @@
 import ImageModeReader from './ImageModeReader.vue';
+export { default as iconMap } from './icons';
 
 export default ImageModeReader;
