+++ packages/reader-named-entities-mode/src/index.js
@@ -1,3 +1,4 @@
-import NamedEntitiesModeReader from "./NamedEntitiesModeReader.vue";
+import NamedEntitiesModeReader from './NamedEntitiesModeReader.vue';
+export { default as iconMap } from './icons';
 
 export default NamedEntitiesModeReader;
