+++ packages/widget-named-entities/package.json
@@ -8,6 +8,7 @@
   "author": "Eldarion",
   "license": "MIT",
   "dependencies": {
+    "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "@scaife-viewer/common": "^0.0.2",
     "@scaife-viewer/entity-map": "^0.0.2",
     "@scaife-viewer/store": "^0.0.4",
