+++ packages/widget-named-entities/src/icons.js
@@ -1,0 +1,17 @@
+import {
+  faUser,
+  faMapMarkerAlt,
+} from '@fortawesome/free-solid-svg-icons';
+
+
+const iconMap = [
+  faUser,
+  faMapMarkerAlt,
+].reduce((map, obj) => {
+  return {
+    ...map,
+    [obj.iconName]: obj,
+  };
+}, {});
+
+export default iconMap;
