+++ packages/widget-named-entities/src/index.js
@@ -1,3 +1,4 @@
 import NamedEntitiesWidget from './NamedEntitiesWidget.vue';
+export { default as iconMap } from './icons';
 
 export default NamedEntitiesWidget;
