+++ src/main.js
@@ -3,23 +3,6 @@ import VueApollo from 'vue-apollo';
 import { sync } from 'vuex-router-sync';
 
 import {
-  faChevronLeft,
-  faChevronDown,
-  faChevronRight,
-  faVolumeUp,
-  faUser,
-  faMapMarkerAlt,
-  faGripLines,
-  faGripLinesVertical,
-  faAlignLeft,
-  faBookOpen,
-  faSearchMinus,
-  faSearchPlus,
-  faHome,
-  faWindowMaximize,
-} from '@fortawesome/free-solid-svg-icons';
-
-import {
   DISPLAY_MODE_SENTENCE_ALIGNMENTS,
   DISPLAY_MODE_FOLIO,
   DISPLAY_MODE_INTERLINEAR,
@@ -37,40 +20,31 @@ import NamedEntitiesModeReader from '@scaife-viewer/reader-named-entities-mode';
 import MetricalModeReader from '@scaife-viewer/reader-metrical-mode';
 import InterlinearModeReader from '@scaife-viewer/reader-interlinear-mode';
 
+import { iconMap as commonIconMap } from "@scaife-viewer/common";
+import { iconMap as audioIconMap } from "@scaife-viewer/widget-audio";
+import { iconMap as namedEntitiesIconMap } from "@scaife-viewer/widget-named-entities";
+import { iconMap as imageModeReaderIconMap } from "@scaife-viewer/reader-image-mode";
+import { iconMap as namedEntitesReaderIconMap } from '@scaife-viewer/reader-named-entities-mode';
+
 import App from './App.vue';
 import store, { apolloProvider } from './store';
 import router from './router';
 
 sync(store, router);
 
-const iconMap = [
-  faChevronLeft,
-  faChevronDown,
-  faChevronRight,
-  faVolumeUp,
-  faUser,
-  faMapMarkerAlt,
-  faGripLines,
-  faGripLinesVertical,
-  faAlignLeft,
-  faBookOpen,
-  faSearchMinus,
-  faSearchPlus,
-  faHome,
-  faWindowMaximize,
-].reduce((map, obj) => {
-  return {
-    ...map,
-    [obj.iconName]: obj,
-  };
-}, {});
-
 Vue.use(SkeletonPlugin, {
-  iconMap,
+  iconMap: {
+    ...commonIconMap,
+    ...audioIconMap,
+    ...namedEntitiesIconMap,
+    ...namedEntitesReaderIconMap,
+    ...imageModeReaderIconMap,
+  },
   config: {
     entityMap: {
-      accessToken: 'pk.eyJ1IjoicGFsdG1hbiIsImEiOiJja2JpNDVpbmUwOGF1MnJwZm91c3VybDVrIn0.KRcXBGtiUWFXkp2uaE5LLw',
-      mapStyle: 'mapbox://styles/paltman/ckbi4thqt156y1ijz5wldui14',
+      accessToken:
+        "pk.eyJ1IjoicGFsdG1hbiIsImEiOiJja2JpNDVpbmUwOGF1MnJwZm91c3VybDVrIn0.KRcXBGtiUWFXkp2uaE5LLw",
+      mapStyle: "mapbox://styles/paltman/ckbi4thqt156y1ijz5wldui14",
     },
     readerComponents: {
       [DISPLAY_MODE_SENTENCE_ALIGNMENTS]: AlignmentsModeReader,
